+++ CHANGELOG.md
@@ -11,10 +11,12 @@ This project <em>does not yet</em> adhere to [Semantic Versioning](https://semve
 
 - Radial annotator: Add pagination to citations slice [#4964]
 - In the TaxonName verbatim author/year task, add a Source selector as an option when citing, and add a modal for assigning page numbers to source-only citations that get added to names
+- Add object ID to Radial annotator/object/navigator header [#4978]
 
 ### Changed
 
 - Upgraded postresql client to version 18
+- Names limit in the Match OTUs to Taxon Names task is now 3k, up from 1k
 
 ### Fixed
 
@@ -22,8 +24,13 @@ This project <em>does not yet</em> adhere to [Semantic Versioning](https://semve
 - Notification sound triggers multiple times when there is more than one notification. [#4964]
 - Error when changing the type of a label from qrcode/barcode to text
 - DwC occurrences importer not checking existing nomenclature date data correctly when matching on author/year
+- Regexes are applied when the user clicks the 'Match' button in the Match OTUs to Taxon Names task
+- "Is absent" checkbox does not display the value saved in session storage when citation form is loaded
+- Include citations on the OTU itself in addition to associated objects, for the endpoint /api/v1/otus/:id/inventory/citations [#4983]
 
 [#4964]: https://github.com/SpeciesFileGroup/taxonworks/issues/4964
+[#4978]: https://github.com/SpeciesFileGroup/taxonworks/issues/4978
+[#4983]: https://github.com/SpeciesFileGroup/taxonworks/issues/4983
 
 ## [0.63.1] - 2026-06-22